+++ CHANGES.md
@@ -1,6 +1,12 @@
 # python-markdown2 Changelog
 
-## python-markdown2 2.4.10 (not yet released)
+## python-markdown2 2.4.11 (not yet released)
+
+- [pull #524] Fix angles being escaped in style blocks (issue #523)
+- [pull #527] Fix base64 images being corrupted in safe mode (issue #526)
+- [pull #529] Add `breaks` extra with ability to hard break on backslashes (issue #525)
+
+## python-markdown2 2.4.10
 
 - [pull #519] Add support for custom extras
 - [pull #519] Drop Python 3.5 support